moving message handling tests to their own files
