+++ Assets/Tests/Runtime/Serialization/MessageHandlerTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12d20767883feb64ebfc388ed9dac57d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
